requirements.txt
CherryPy
dbus-python
flake8
hacking
Jinja2
netifaces
PyGObject
